% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bi_var.R
\name{plot_bi_var}
\alias{plot_bi_var}
\title{Plot bivariate correlations from bi_var()}
\usage{
plot_bi_var(x)
}
\arguments{
\item{x}{Result object returned by \code{bi_var()}}
}
\value{
A ggplot object
}
\description{
Creates a horizontal bar chart of Pearson correlations returned by \code{bi_var()}.
}
