% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cfa_var.R
\name{plot_cfa_var}
\alias{plot_cfa_var}
\title{Plot standardized CFA loadings}
\usage{
plot_cfa_var(x)
}
\arguments{
\item{x}{Result object returned by \code{cfa_var()}}
}
\value{
A ggplot object
}
\description{
Creates a bar chart of standardized loadings from the result
returned by \code{cfa_var()}.
}
