% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contrib_balance.R
\name{plot_contrib_balance}
\alias{plot_contrib_balance}
\title{Plot contribution balance}
\usage{
plot_contrib_balance(x)
}
\arguments{
\item{x}{Result object returned from \code{contrib_balance()}}
}
\value{
A ggplot object
}
\description{
Creates a horizontal bar chart of variable contribution shares from
the \code{contrib_balance()} function.
}
