% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factor_var.R
\name{plot_factor_var}
\alias{plot_factor_var}
\title{Plot cumulative variance explained from factor_var()}
\usage{
plot_factor_var(x)
}
\arguments{
\item{x}{Result object returned by \code{factor_var()}}
}
\value{
A ggplot object
}
\description{
Creates a plot of cumulative variance explained by PCA components
using the result returned by \code{factor_var()}.
}
